+++ apps/scouting/src/features/scouting-schedule/components/ScoutingScheduleBoard.tsx
@@ -6,12 +6,9 @@ import {
   type DragEndEvent,
   PointerSensor,
   TouchSensor,
-  useDraggable,
-  useDroppable,
   useSensor,
   useSensors,
 } from "@dnd-kit/core";
-import { CSS } from "@dnd-kit/utilities";
 import { Avatar, AvatarFallback, AvatarImage } from "@repo/ui/components/avatar";
 import { Badge } from "@repo/ui/components/badge";
 import { Button } from "@repo/ui/components/button";
@@ -32,7 +29,6 @@ import {
 } from "@repo/ui/components/command";
 import { Input } from "@repo/ui/components/input";
 import { Label } from "@repo/ui/components/label";
-import { ScrollArea } from "@repo/ui/components/scroll-area";
 import {
   Select,
   SelectContent,
@@ -50,6 +46,7 @@ import {
   TableRow,
 } from "@repo/ui/components/table";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@repo/ui/components/tabs";
+import { TypographyH1 } from "@repo/ui/components/typography";
 import { cn } from "@repo/ui/lib/utils";
 import {
   ArrowLeftIcon,
@@ -57,8 +54,6 @@ import {
   PlusIcon,
   SaveIcon,
   Trash2Icon,
-  UsersIcon,
-  XIcon,
 } from "lucide-react";
 import Link from "next/link";
 import { useEffect, useMemo, useState, useTransition } from "react";
@@ -69,10 +64,7 @@ import {
   formatScheduleDateRange,
   groupMatchesByShifts,
 } from "../logic";
-import type {
-  ScoutingScheduleBoardMatch,
-  ScoutingScheduleBoardTeamSlot,
-} from "../logic";
+import type { ScoutingScheduleBoardMatch } from "../logic";
 import type {
   ScoutingScheduleDetailData,
   ScoutingScheduleMember,
@@ -82,6 +74,7 @@ import type {
   ScoutingScheduleInfoSection,
   ScoutingScheduleShiftSection,
 } from "../types";
+import { RosterSidebar, ShiftGroup } from "./ScoutingScheduleBoardSections";
 
 interface ScoutingScheduleBoardProps {
   data: ScoutingScheduleDetailData;
@@ -158,170 +151,6 @@ function ScheduleMemberAvatar({
   );
 }
 
-function DraggableRosterMember({ member }: { member: ScoutingScheduleMember }) {
-  const { attributes, listeners, setNodeRef, transform, isDragging } = useDraggable({
-    id: `member:${member.id}`,
-    data: { memberId: member.id },
-  });
-
-  return (
-    <button
-      ref={setNodeRef}
-      type="button"
-      style={{
-        transform: CSS.Translate.toString(transform),
-      }}
-      className={cn(
-        "flex w-full min-w-0 items-center gap-3 rounded-md border bg-card px-3 py-2 text-left transition-colors hover:bg-accent/20",
-        isDragging && "opacity-60 shadow-lg ring-2 ring-primary/30"
-      )}
-      {...attributes}
-      {...listeners}
-    >
-      <ScheduleMemberAvatar member={member} className="shrink-0" />
-      <div className="min-w-0 flex-1 overflow-hidden">
-        <p className="truncate text-sm font-medium">{member.name}</p>
-        <p className="truncate text-[11px] text-muted-foreground">{roleLabel(member.role)}</p>
-      </div>
-      <Badge variant="outline" className="shrink-0 whitespace-nowrap px-2.5">
-        Drag
-      </Badge>
-    </button>
-  );
-}
-
-function CompactRosterMember({ member }: { member: ScoutingScheduleMember }) {
-  return (
-    <div className="flex min-w-0 items-center gap-2 rounded-md border bg-card px-2.5 py-2">
-      <ScheduleMemberAvatar member={member} className="size-6 shrink-0" />
-      <div className="min-w-0 flex-1 overflow-hidden">
-        <p className="truncate text-sm font-medium">{member.name}</p>
-        <p className="truncate text-[11px] text-muted-foreground">{roleLabel(member.role)}</p>
-      </div>
-    </div>
-  );
-}
-
-function CoordinatorDropZone({
-  shiftId,
-  coordinator,
-  canEdit,
-  onOpenPicker,
-}: {
-  shiftId: string;
-  coordinator: ScoutingScheduleMember | null;
-  canEdit: boolean;
-  onOpenPicker: (target: PickerTarget) => void;
-}) {
-  const { isOver, setNodeRef } = useDroppable({
-    id: `coordinator:${shiftId}`,
-    disabled: !canEdit,
-  });
-
-  return (
-    <button
-      ref={setNodeRef}
-      type="button"
-      onClick={() => canEdit && onOpenPicker({ kind: "coordinator", shiftId })}
-      disabled={!canEdit}
-      className={cn(
-        "flex min-h-9 w-full items-center gap-2 rounded-sm border px-2 py-1.5 text-left transition-colors",
-        "border-border bg-background hover:bg-accent/10",
-        isOver && "border-primary/60 bg-accent/10",
-        !canEdit && "cursor-default"
-      )}
-    >
-      <span className="rounded-sm bg-muted px-1.5 py-0.5 text-[10px] font-semibold text-muted-foreground">
-        EC
-      </span>
-      <span className="truncate text-sm font-medium">
-        {coordinator?.name ?? "Assign event coordinator"}
-      </span>
-    </button>
-  );
-}
-
-function AssignmentSeat({
-  teamMatchId,
-  slotIndex,
-  member,
-  canEdit,
-  isPending,
-  onOpenPicker,
-  onClear,
-}: {
-  teamMatchId: number;
-  slotIndex: 1 | 2;
-  member: ScoutingScheduleMember | null;
-  canEdit: boolean;
-  isPending: boolean;
-  onOpenPicker: (target: PickerTarget) => void;
-  onClear: (target: PickerTarget) => void;
-}) {
-  const { isOver, setNodeRef } = useDroppable({
-    id: `seat:${teamMatchId}:${slotIndex}`,
-    disabled: !canEdit,
-  });
-
-  return (
-    <div className="relative">
-      <button
-        ref={setNodeRef}
-        type="button"
-        disabled={!canEdit || isPending}
-        onClick={() => canEdit && onOpenPicker({ kind: "seat", teamMatchId, slotIndex })}
-        className={cn(
-          "flex min-h-8 w-full items-center gap-2 rounded-sm border px-2 py-1.5 text-left transition-colors",
-          member
-            ? "border-border bg-background hover:bg-accent/10"
-            : "border-border border-dashed bg-background hover:bg-accent/10",
-          isOver && "border-primary/60 bg-accent/10",
-          !canEdit && "cursor-default border-solid"
-        )}
-      >
-        <span className="rounded-sm bg-muted px-1.5 py-0.5 text-[10px] font-semibold tabular-nums text-muted-foreground">
-          S{slotIndex}
-        </span>
-        {member ? (
-          <div className="flex min-w-0 items-center gap-2">
-            <ScheduleMemberAvatar member={member} className="size-6" />
-            <div className="min-w-0">
-              <p className="truncate text-sm font-medium">{member.name}</p>
-              <p className="truncate text-[10px] text-muted-foreground">{roleLabel(member.role)}</p>
-            </div>
-          </div>
-        ) : (
-          <div className="min-w-0">
-            <p className="text-sm font-medium">{canEdit ? "Assign scout" : "Unassigned"}</p>
-          </div>
-        )}
-      </button>
-
-      {member && canEdit && (
-        <Button
-          type="button"
-          variant="ghost"
-          size="icon-xs"
-          className="absolute top-2 right-2"
-          disabled={isPending}
-          onClick={() => onClear({ kind: "seat", teamMatchId, slotIndex })}
-        >
-          <XIcon className="size-3" />
-        </Button>
-      )}
-    </div>
-  );
-}
-
-const TEAM_COLUMNS = [
-  { alliance: "red", position: 1, label: "Red 1" },
-  { alliance: "red", position: 2, label: "Red 2" },
-  { alliance: "red", position: 3, label: "Red 3" },
-  { alliance: "blue", position: 1, label: "Blue 1" },
-  { alliance: "blue", position: 2, label: "Blue 2" },
-  { alliance: "blue", position: 3, label: "Blue 3" },
-] as const;
-
 function shouldAutoSplitShiftSections(
   matches: ScoutingScheduleBoardMatch[],
   shifts: ScoutingScheduleShiftSection[]
@@ -386,85 +215,19 @@ function ScheduleInfoPreview({ sections }: { sections: ScoutingScheduleInfoSecti
   );
 }
 
-function findTeamSlot(
-  matchEntry: ScoutingScheduleBoardMatch,
-  alliance: "red" | "blue",
-  position: 1 | 2 | 3
-) {
-  return (
-    matchEntry.teams.find((teamSlot) => teamSlot.alliance === alliance && teamSlot.position === position) ??
-    null
-  );
-}
-
-function TeamAssignmentCell({
-  teamSlot,
-  canEdit,
-  memberMap,
-  pendingAssignmentKey,
-  onOpenPicker,
-  onAssignMember,
-}: {
-  teamSlot: ScoutingScheduleBoardTeamSlot | null;
-  canEdit: boolean;
-  memberMap: Map<string, ScoutingScheduleMember>;
-  pendingAssignmentKey: string | null;
-  onOpenPicker: (target: PickerTarget) => void;
-  onAssignMember: (target: PickerTarget, memberId: string | null) => void;
-}) {
-  if (!teamSlot) {
-    return <div className="py-3 text-sm text-muted-foreground">Unavailable</div>;
-  }
-
-  const teamNumberClasses =
-    teamSlot.alliance === "red" ? "text-red-600 dark:text-red-400" : "text-blue-600 dark:text-blue-400";
-
-  return (
-    <div className="space-y-2">
-      <div className="space-y-0.5">
-        <p className={cn("font-mono text-xl font-bold leading-none", teamNumberClasses)}>
-          {teamSlot.teamNumber}
-        </p>
-        <p className="truncate text-[13px] text-muted-foreground">{teamSlot.teamName}</p>
-      </div>
-
-      <div className="space-y-1">
-        {teamSlot.assignments.map((assignment) => {
-          const assignedMember = assignment.memberId
-            ? memberMap.get(assignment.memberId) ?? null
-            : null;
-          const seatKey = `${teamSlot.teamMatchId}:${assignment.slotIndex}`;
-
-          return (
-            <AssignmentSeat
-              key={seatKey}
-              teamMatchId={teamSlot.teamMatchId}
-              slotIndex={assignment.slotIndex}
-              member={assignedMember}
-              canEdit={canEdit}
-              isPending={pendingAssignmentKey === seatKey}
-              onOpenPicker={onOpenPicker}
-              onClear={(target) => onAssignMember(target, null)}
-            />
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
 export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardProps) {
   const [schedule, setSchedule] = useState(data.schedule);
   const [matches, setMatches] = useState(data.matches);
   const [detailsMode, setDetailsMode] = useState(false);
   const [memberSearch, setMemberSearch] = useState("");
   const [pickerTarget, setPickerTarget] = useState<PickerTarget | null>(null);
-  const [pendingAssignmentKey, setPendingAssignmentKey] = useState<string | null>(null);
+  const [activeAssignmentKey, setActiveAssignmentKey] = useState<string | null>(null);
   const [savedSchedule, setSavedSchedule] = useState(data.schedule);
   const [savedDetailsSnapshot, setSavedDetailsSnapshot] = useState(() =>
     snapshotScheduleDetails(data.schedule)
   );
   const [isSavingDetails, startSavingDetails] = useTransition();
+  const [isSavingAssignment, startSavingAssignment] = useTransition();
 
   const sensors = useSensors(useSensor(PointerSensor), useSensor(TouchSensor));
 
@@ -494,6 +257,7 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
   );
 
   const detailsAreDirty = snapshotScheduleDetails(schedule) !== savedDetailsSnapshot;
+  const pendingAssignmentKey = isSavingAssignment ? activeAssignmentKey : null;
   const pickerTitle =
     pickerTarget?.kind === "coordinator"
       ? "Assign event coordinator"
@@ -544,26 +308,26 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
     );
   }
 
-  async function assignMember(teamMatchId: number, slotIndex: 1 | 2, memberId: string | null) {
+  function assignMember(teamMatchId: number, slotIndex: 1 | 2, memberId: string | null) {
     const seatKey = `${teamMatchId}:${slotIndex}`;
-    setPendingAssignmentKey(seatKey);
+    setActiveAssignmentKey(seatKey);
 
-    const result = await updateScoutingScheduleAssignment({
-      scheduleId: schedule.id,
-      teamMatchId,
-      slotIndex,
-      memberId,
+    startSavingAssignment(async () => {
+      const result = await updateScoutingScheduleAssignment({
+        scheduleId: schedule.id,
+        teamMatchId,
+        slotIndex,
+        memberId,
+      });
+
+      if (result.error) {
+        toast.error(result.error);
+        return;
+      }
+
+      updateSeatState(teamMatchId, slotIndex, memberId);
+      setPickerTarget(null);
     });
-
-    setPendingAssignmentKey(null);
-
-    if (result.error) {
-      toast.error(result.error);
-      return;
-    }
-
-    updateSeatState(teamMatchId, slotIndex, memberId);
-    setPickerTarget(null);
   }
 
   function assignCoordinator(shiftId: string, memberId: string | null) {
@@ -578,16 +342,18 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
   function handleDragEnd(event: DragEndEvent) {
     if (!canEdit || !event.over) return;
 
-    const memberId = String(event.active.id).replace(/^member:/, "");
-    const [prefix, idA, idB] = String(event.over.id).split(":");
+    const memberId = event.active.data.current?.memberId;
+    const drop = event.over.data.current as PickerTarget | undefined;
 
-    if (prefix === "seat" && idA && idB) {
-      void assignMember(Number(idA), Number(idB) as 1 | 2, memberId);
+    if (typeof memberId !== "string" || !drop) return;
+
+    if (drop.kind === "seat" && drop.teamMatchId != null && drop.slotIndex != null) {
+      assignMember(drop.teamMatchId, drop.slotIndex, memberId);
       return;
     }
 
-    if (prefix === "coordinator" && idA) {
-      assignCoordinator(idA, memberId);
+    if (drop.kind === "coordinator" && drop.shiftId) {
+      assignCoordinator(drop.shiftId, memberId);
     }
   }
 
@@ -698,7 +464,7 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
       return;
     }
 
-    void assignMember(target.teamMatchId, target.slotIndex, memberId);
+    assignMember(target.teamMatchId, target.slotIndex, memberId);
   }
 
   return (
@@ -718,7 +484,7 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
               <Badge variant="outline">{data.schedule.event.eventCode.toUpperCase()}</Badge>
             </div>
             <div>
-              <h1 className="text-3xl tracking-tight">{schedule.name}</h1>
+              <TypographyH1>{schedule.name}</TypographyH1>
               <p className="mt-2 text-muted-foreground">
                 {data.schedule.event.name} •{" "}
                 {formatScheduleDateRange(
@@ -1064,44 +830,12 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
             "xl:grid-cols-[248px_minmax(0,1fr)]"
           )}
         >
-          <aside className="order-2 space-y-4 xl:order-1">
-              <Card className="xl:sticky xl:top-24">
-                <CardHeader>
-                  <CardTitle className="flex items-center gap-2">
-                    <UsersIcon className="size-4" />
-                    Active roster
-                  </CardTitle>
-                  <CardDescription>
-                    {canEdit
-                      ? "Drag scouts into the table or click a seat to assign."
-                      : "Current organization members for this event schedule."}
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="space-y-4">
-                  <Input
-                    value={memberSearch}
-                    onChange={(event) => setMemberSearch(event.target.value)}
-                    placeholder="Search members"
-                  />
-                  <ScrollArea className="h-[12rem] xl:h-[calc(100vh-18rem)]">
-                    <div className="space-y-2 pr-3">
-                      {filteredMembers.map((member) =>
-                        canEdit ? (
-                          <DraggableRosterMember key={member.id} member={member} />
-                        ) : (
-                          <CompactRosterMember key={member.id} member={member} />
-                        )
-                      )}
-                      {filteredMembers.length === 0 && (
-                        <p className="rounded-xl border border-dashed px-4 py-6 text-center text-sm text-muted-foreground">
-                          No members match that search.
-                        </p>
-                      )}
-                    </div>
-                  </ScrollArea>
-                </CardContent>
-              </Card>
-            </aside>
+          <RosterSidebar
+            canEdit={canEdit}
+            filteredMembers={filteredMembers}
+            memberSearch={memberSearch}
+            onMemberSearchChange={setMemberSearch}
+          />
 
           <section className="order-1 min-w-0 space-y-8 xl:order-2">
             {matches.length === 0 ? (
@@ -1115,101 +849,15 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
               </Card>
             ) : (
               groupedMatches.map((group) => (
-                <section key={group.id} className="space-y-3">
-                  <div className="overflow-hidden rounded-xl border">
-                    <div className="overflow-x-auto">
-                      <div className="min-w-[1180px] border-b bg-card px-4 py-3">
-                        <div className="flex flex-wrap items-center gap-4">
-                          <h2 className="text-base font-semibold text-foreground">{group.label}</h2>
-                          {group.id !== "ungrouped" ? (
-                            <div className="w-full max-w-xs sm:w-[280px]">
-                              <CoordinatorDropZone
-                                shiftId={group.id}
-                                coordinator={
-                                  group.coordinatorMemberId
-                                    ? memberMap.get(group.coordinatorMemberId) ?? null
-                                    : null
-                                }
-                                canEdit={canEdit}
-                                onOpenPicker={setPickerTarget}
-                              />
-                            </div>
-                          ) : null}
-                        </div>
-                      </div>
-                      <Table className="min-w-[1180px] table-fixed">
-                        <TableHeader>
-                          <TableRow>
-                            <TableHead className="w-20 bg-card text-xs font-semibold uppercase tracking-[0.14em]">
-                              Match
-                            </TableHead>
-                            {TEAM_COLUMNS.map((column) => (
-                              <TableHead
-                                key={`${column.alliance}-${column.position}`}
-                                className={cn(
-                                  "w-[176px] bg-card text-left text-xs font-semibold uppercase tracking-[0.14em]",
-                                  column.alliance === "blue" && column.position === 1 && "border-l-4 border-muted",
-                                  column.alliance === "red"
-                                    ? "text-red-600 dark:text-red-400"
-                                    : "text-blue-600 dark:text-blue-400"
-                                )}
-                              >
-                                {column.label}
-                              </TableHead>
-                            ))}
-                          </TableRow>
-                        </TableHeader>
-                        <TableBody>
-                          {group.matches.map((matchEntry, index) => (
-                            <TableRow
-                              key={matchEntry.matchNumber}
-                              className="align-top hover:bg-muted/20"
-                            >
-                              <TableCell
-                                className={cn(
-                                  "border-r bg-card px-3 py-3 align-top",
-                                  index % 2 === 1 && "bg-muted/5"
-                                )}
-                              >
-                                <div className="space-y-1">
-                                  <p className="font-mono text-lg font-semibold">
-                                    {matchEntry.matchNumber}
-                                  </p>
-                                  <p className="text-[11px] uppercase tracking-[0.14em] text-muted-foreground">
-                                    QM
-                                  </p>
-                                </div>
-                              </TableCell>
-
-                              {TEAM_COLUMNS.map((column) => (
-                                <TableCell
-                                  key={`${matchEntry.matchNumber}-${column.alliance}-${column.position}`}
-                                  className={cn(
-                                    "bg-card px-3 py-3 align-top",
-                                    column.alliance === "blue" && column.position === 1 && "border-l-4 border-muted"
-                                  )}
-                                >
-                                  <TeamAssignmentCell
-                                    teamSlot={findTeamSlot(
-                                      matchEntry,
-                                      column.alliance,
-                                      column.position
-                                    )}
-                                    canEdit={canEdit}
-                                    memberMap={memberMap}
-                                    pendingAssignmentKey={pendingAssignmentKey}
-                                    onOpenPicker={setPickerTarget}
-                                    onAssignMember={handleSeatAssignment}
-                                  />
-                                </TableCell>
-                              ))}
-                            </TableRow>
-                          ))}
-                        </TableBody>
-                      </Table>
-                    </div>
-                  </div>
-                </section>
+                <ShiftGroup
+                  key={group.id}
+                  group={group}
+                  canEdit={canEdit}
+                  memberMap={memberMap}
+                  pendingAssignmentKey={pendingAssignmentKey}
+                  onOpenPicker={setPickerTarget}
+                  onAssignMember={handleSeatAssignment}
+                />
               ))
             )}
           </section>
@@ -1243,7 +891,7 @@ export function ScoutingScheduleBoard({ data, canEdit }: ScoutingScheduleBoardPr
                   }
 
                   if (pickerTarget.kind === "seat" && pickerTarget.teamMatchId && pickerTarget.slotIndex) {
-                    void assignMember(pickerTarget.teamMatchId, pickerTarget.slotIndex, member.id);
+                    assignMember(pickerTarget.teamMatchId, pickerTarget.slotIndex, member.id);
                   }
                 }}
               >
