+++ apps/scouting/src/features/scouting/drive-ranking/components/DriveRatingHistory.tsx
@@ -144,5 +144,3 @@ export function DriveRatingHistoryChart({ history }: { history: DriveRatingHisto
     </Card>
   );
 }
-
-//comment to fix build, just so I can push
